code refactoring and default, type for platforms added
